Keep output image in general layout
Layout transition has a cost (assuming concept exists in the actual
hardware. It does actual transformation of the image data/metadata, and
even if it is accelerated by the hardware I assume it is not free.

In the real-world applications where the program does  multiple draws
it is benefitial to do the layout trasition - its costs amortizes over
multiple draws in the form of a saved memory bandwidth.

This simple application does a single draw call to render the model.
Jumping between memory layouts few times per frame might actually
slow down things, if the bandwidth savings from a single draw call do
not cover the time spend on performing at least two layout transitions.

As a bonus it gets easier to reason about the code (the same layout,
not much to talk about).

If you ask me, but why not to make layout transition as a good practice,
because this code might be used as a foundation to create more advanced
application. The problem is that we don't know exactly what the more
advanced application will do. What happens in practice is that
additional generalization does not guess correctly what is good for
that future application, but even worse - we fail to write good
solution for this specific application. Another goals of me writing
this code (mostly for myself) is not only Vulkan (which is just an API
for what it worse), but to practise writing good code.